move filtered_params method to helper and add only_path: true when merging parameters
