+++ app/controllers/concerns/enju_library/controller.rb
@@ -8,7 +8,6 @@ module EnjuLibrary
       rescue_from Pundit::NotAuthorizedError, with: :render_403
       # rescue_from ActiveRecord::RecordNotFound, with: :render_404
       # rescue_from ActionView::MissingTemplate, with: :render_404_invalid_format
-      helper_method :filtered_params
     end
 
     private
@@ -244,10 +243,6 @@ module EnjuLibrary
         @subscription = Subscription.find(params[:subscription_id])
         authorize @subscription, :show?
       end
-    end
-
-    def filtered_params
-      params.permit([:q, :query, :view, :format, :order, :sort_by, :page, :per_page])
     end
 
     class InvalidLocaleError < StandardError
