+++ app/controllers/messages_controller.rb
@@ -166,10 +166,4 @@ class MessagesController < ApplicationController
       end
     end
   end
-
-  def filtered_params
-    params.permit([:view, :format, :page, :mode, :sort_by, :per_page])
-  end
-
-  helper_method :filtered_params
 end
