+++ app/helpers/checkouts_helper.rb
@@ -1,0 +1,5 @@
+module CheckoutsHelper
+  def filtered_params
+    params.permit([:user_id, :days_overdue, :reserved])
+  end
+end
