+++ app/helpers/events_helper.rb
@@ -1,0 +1,5 @@
+module EventsHelper
+  def filtered_params
+    params.permit([:format, :library_id, :mode, :query, :page, :per_page])
+  end
+end
