+++ app/helpers/libraries_helper.rb
@@ -1,0 +1,5 @@
+module LibrariessHelper
+  def filtered_params
+    params.permit([:q, :query, :view, :format, :order, :sort_by, :page, :per_page])
+  end
+end
