+++ app/helpers/messages_helper.rb
@@ -1,2 +1,5 @@
 module MessagesHelper
+  def filtered_params
+    params.permit([:view, :format, :page, :mode, :sort_by, :per_page])
+  end
 end
