+++ app/views/checkouts/_index_overdue.html.erb
@@ -27,12 +27,12 @@
   <h4><%= t('page.other_format') %></h4>
   <ul>
     <li>
-      <%= link_to 'RSS', url_for(filtered_params.merge(format: :rss, days_overdue: params[:days_overdue])) -%>
-      <%= link_to image_tag('icons/feed.png', size: '16x16', class: 'enju_icon', alt: 'RSS'), url_for(filtered_params.merge(format: :rss, days_overdue: params[:days_overdue])) -%>
+      <%= link_to 'RSS', url_for(filtered_params.merge(only_path: true, format: :rss, days_overdue: params[:days_overdue])) -%>
+      <%= link_to image_tag('icons/feed.png', size: '16x16', class: 'enju_icon', alt: 'RSS'), url_for(filtered_params.merge(only_path: true, format: :rss, days_overdue: params[:days_overdue])) -%>
     </li>
     <li>
-      <%= link_to 'TSV', url_for(filtered_params.merge(format: :txt, days_overdue: params[:days_overdue], locale: @locale.to_s)) -%>
-      <%= link_to image_tag('icons/page_white_excel.png', size: '16x16', class: 'enju_icon', alt: 'TSV'), url_for(filtered_params.merge(format: :txt, days_overdue: params[:days_overdue], locale: @locale.to_s)) -%>
+      <%= link_to 'TSV', url_for(filtered_params.merge(only_path: true, format: :txt, days_overdue: params[:days_overdue], locale: @locale.to_s)) -%>
+      <%= link_to image_tag('icons/page_white_excel.png', size: '16x16', class: 'enju_icon', alt: 'TSV'), url_for(filtered_params.merge(only_path: true, format: :txt, days_overdue: params[:days_overdue], locale: @locale.to_s)) -%>
     </li>
   </ul>
 </div>
