+++ app/views/reserves/index.html.erb
@@ -103,15 +103,15 @@
 
   <p>
     <%- if @user -%>
-      <%= link_to image_tag('icons/feed.png', size: '16x16', class: 'enju_icon', alt: t('page.feed')), url_for(filtered_params.merge(user_id: @user.username, format: :rss)) -%>
-      (<%= link_to 'RSS', url_for(filtered_params.merge(user_id: @user.username, format: :rss)) -%>)
-      <%= link_to image_tag('icons/page_white_excel.png', size: '16x16', class: 'enju_icon', alt: 'TSV'), url_for(filtered_params.merge(user_id: @user.username, format: :txt, locale: @locale.to_s)) -%>
-      (<%= link_to 'TSV', url_for(filtered_params.merge(user_id: @user.username, format: :txt, locale: @locale.to_s)) -%>)
+      <%= link_to image_tag('icons/feed.png', size: '16x16', class: 'enju_icon', alt: t('page.feed')), url_for(filtered_params.merge(only_path: true, user_id: @user.username, format: :rss)) -%>
+      (<%= link_to 'RSS', url_for(filtered_params.merge(only_path: true, user_id: @user.username, format: :rss)) -%>)
+      <%= link_to image_tag('icons/page_white_excel.png', size: '16x16', class: 'enju_icon', alt: 'TSV'), url_for(filtered_params.merge(only_path: true, user_id: @user.username, format: :txt, locale: @locale.to_s)) -%>
+      (<%= link_to 'TSV', url_for(filtered_params.merge(only_path: true, user_id: @user.username, format: :txt, locale: @locale.to_s)) -%>)
     <%- else -%>
-      <%= link_to image_tag('icons/feed.png', size: '16x16', class: 'enju_icon', alt: t('page.feed')), url_for(filtered_params.merge(format: :rss)) -%>
-      (<%= link_to 'RSS', url_for(filtered_params.merge(format: :rss)) -%>)
-      <%= link_to image_tag('icons/page_white_excel.png', size: '16x16', class: 'enju_icon', alt: 'TSV'), url_for(filtered_params.merge(format: :txt, locale: @locale.to_s)) -%>
-      (<%= link_to 'TSV', url_for(filtered_params.merge(format: :txt, locale: @locale.to_s)) -%>)
+      <%= link_to image_tag('icons/feed.png', size: '16x16', class: 'enju_icon', alt: t('page.feed')), url_for(filtered_params.merge(only_path: true, format: :rss)) -%>
+      (<%= link_to 'RSS', url_for(filtered_params.merge(only_path: true, format: :rss)) -%>)
+      <%= link_to image_tag('icons/page_white_excel.png', size: '16x16', class: 'enju_icon', alt: 'TSV'), url_for(filtered_params.merge(only_path: true, format: :txt, locale: @locale.to_s)) -%>
+      (<%= link_to 'TSV', url_for(filtered_params.merge(only_path: true, format: :txt, locale: @locale.to_s)) -%>)
     <%- end -%>
   </p>
 </div>
